// src/model/smartEnum.ts
export enum VENDOR {
    DEFAULT = 'Smart', // HDD
    CORSAIR = 'SmartCorsair',
    INDLINX = 'SmartIndlinx',
    INTEL = 'SmartIntel',
    J_MICRON60X = 'SmartJMicron60x',
    J_MICRON61X = 'SmartJMicron61x',
    KINGSTON = 'SmartKingston',
    MICRON = 'SmartMicron',
    MICRON_MU02 = 'SmartMicronMU02',
    MTRON = 'SmartMtron',
    NVM_E = 'SmartNVMe',
    OCZ = 'SmartOcz',
    OCZ_VECTOR = 'SmartOczVector',
    PLEXTOR = 'SmartPlextor',
    SAMSUNG = 'SmartSamsung',
    SAND_FORCE = 'SmartSandForce',
    SANDISK = 'SmartSanDisk',
    SANDISK_GB = 'SmartSanDiskGb',
    GENERAL = 'SmartSsd',
    TOSHIBA = 'SmartToshiba'
}

export enum STATUS {
    SUCCESS,
    WARN,
    DANGER
}

export enum SANDFORCE {
    READ_ERROR_RATE = 1
}